🐛fix：createTime filling bug
