+++ pojo/src/main/java/com/mingri/dto/user/SysUserDTO.java
@@ -1,4 +1,4 @@
-package com.mingri.dto;
+package com.mingri.dto.user;
 
 import lombok.Data;
 
@@ -6,8 +6,6 @@ import java.io.Serializable;
 
 @Data
 public class SysUserDTO implements Serializable {
-
-    private Long id;
 
     private String nickName;
