+++ server/src/main/java/com/mingri/service/ISysUserService.java
@@ -1,7 +1,7 @@
 package com.mingri.service;
 
-import com.mingri.dto.SysUserLoginDTO;
-import com.mingri.dto.SysUserRegisterDTO;
+import com.mingri.dto.user.SysUserLoginDTO;
+import com.mingri.dto.user.SysUserRegisterDTO;
 import com.mingri.entity.LoginUser;
 import com.mingri.entity.SysUser;
 import com.baomidou.mybatisplus.extension.service.IService;
